+++ CHANGELOG.md
@@ -3,6 +3,7 @@
 ## 0.1.1
 
 - `PerfectShape::QuadraticBezierCurve` (two end points and one control point)
+- `PerfectShape::Path` (having quadratic bezier curves in addition to points and lines)
 
 ## 0.1.0